Move Spanned into source module
