+++ fathom/src/surface/elaboration/unification.rs
@@ -20,10 +20,11 @@ use std::sync::Arc;
 
 use crate::alloc::SliceVec;
 use crate::core::semantics::{
-    self, ArcValue, Closure, Elim, Head, Spanned, SplitBranches, Telescope, Value,
+    self, ArcValue, Closure, Elim, Head, SplitBranches, Telescope, Value,
 };
 use crate::core::{Prim, Term};
 use crate::env::{EnvLen, GlobalVar, LocalVar, SharedEnv, SliceEnv, UniqueEnv};
+use crate::source::Spanned;
 use crate::StringId;
 
 /// Errors encountered during unification.
